fix: OI change is a float64, not an int64 as SPEC 3.4 claims
Caught on the device: the detail screen rendered
+4581235513960227840.00%. Offset 139 read as an int64 yields ~4.6e18;
read as a float64 it is -5.82 to +2.11 across both recorded sessions,
which is a plausible OI change. Confirmed against two independent
captures before changing the decoder.

No test caught this because every assertion aimed at that field was
written from the same wrong premise as the decoder. A round-trip test
cannot find an error like that -- only a plausibility test can, so the
suite now range-checks every decoded double, asserts day OHLC is
internally consistent, and asserts circuit limits bracket the price.
Reverting the fix turns two tests red.

The OHLC check skips contracts with zero high: an illiquid strike that
has not traded today reports zero OHLC while still carrying an LTP from
an earlier session. That is real data, not a decode error.

Also routes Log through debugPrint in debug builds, because
developer.log reaches the VM service but not logcat -- and SPEC 6
criterion 2 asks for back-navigation to be proved with a log line.

+++ lib/data/broker/tick_decoder.dart
@@ -169,10 +169,21 @@ MarketTick decodeSnapQuote(Uint8List bytes) {
       _Offsets.totalSellQuantity,
       Endian.little,
     ),
-    // A percentage, not a price: no ÷100.
-    openInterestChangePercent: data
-        .getInt64(_Offsets.openInterestChangePercent, Endian.little)
-        .toDouble(),
+    // **A float64, despite §3.4's table saying int64.** Read as an integer
+    // this yields 4.58e18 — astronomical nonsense that renders as
+    // "+4581235513960227840.00%". As a float64 it reads -5.82 to +2.11 across
+    // both recorded sessions, which is a plausible OI change.
+    //
+    // Caught on the device rather than by a test, because every assertion
+    // pointed at this field was written from the same wrong premise as the
+    // decoder. The lesson is that a *decoded value* needs a plausibility check
+    // and not only a round-trip — the spec's own table was the thing at fault.
+    //
+    // Already a percentage, so no ÷100.
+    openInterestChangePercent: data.getFloat64(
+      _Offsets.openInterestChangePercent,
+      Endian.little,
+    ),
     upperCircuit: _rupees(data, _Offsets.upperCircuit),
     lowerCircuit: _rupees(data, _Offsets.lowerCircuit),
     fiftyTwoWeekHigh: _rupees(data, _Offsets.fiftyTwoWeekHigh),
